replaced std classes for structures and arrays

+++ headers/kernels.h
@@ -3,10 +3,10 @@
 #define KERNELS_H
 
 #include <cuda_runtime.h>
-#include "common.h"
+#include "types.h"
 
 
-__global__ void conflictMatrixKernel(int *conflictMatrix, int *adjMatrix, int *chromosome, unsigned int n); 
+__global__ void conflictMatrixKernel(ushort *conflictMatrix, ushort *adjMatrix, ushort *chromosome, unsigned int n); 
 std::vector<std::vector<int>*> *crossover(std::vector<int> *first, std::vector<int> *second, int n);
